Look for FamilyLocal if inet or inet6 address is loopback
libxcb uses FamilyLocal authorization if the host name or IP in the
display string is from the loopback device. This patch adds the same
behavior to xauth.
This fixes a long standing problem that for ssh tunneled connections
a display variable of the form: localhost:<N>.<M> leads to correct
authorization when an X client is started but "xauth list $DISPLAY"
returns nothing.

Signed-off-by: Egbert Eich <eich@freedesktop.org>