Fixed missing EXTRA_DIST in tests. Extended README for releasing.
The tests/Makefile.am missed the EXTRA_DIST definition which resulted in
missing files for the test done with "make check". This was fixed.
README was extended to describe howto release a new version of this
software and to clarify something from ReleaseHOWTO of x.org.
Prepare anything for releasing 1.0.9.

Signed-off-by: Dr. Tilmann Bubeck <tilmann@bubecks.de>